Unit testing updates
