+++ ppr-ui/src/components/collateral/general/GenColAmend.vue
