+++ ppr-ui/src/components/collateral/general/GenColEdit.vue
