+++ ppr-ui/src/components/collateral/general/GenColSummary.vue
