+++ ppr-ui/src/components/collateral/general/GeneralCollateral.vue
