+++ ppr-ui/src/components/collateral/general/factories/useGeneralCollateral.ts
