+++ ppr-ui/src/components/collateral/general/index.ts
