+++ ppr-ui/src/components/collateral/vehicle/EditCollateral.vue
