+++ ppr-ui/src/components/collateral/vehicle/VehicleCollateral.vue
