+++ ppr-ui/src/components/collateral/vehicle/factories/collateralFormValidator.ts
