+++ ppr-ui/src/components/collateral/vehicle/factories/index.ts
