+++ ppr-ui/src/components/collateral/vehicle/factories/useCollateralValidation.ts
