+++ ppr-ui/src/components/collateral/vehicle/factories/useVehicle.ts
