+++ ppr-ui/src/components/collateral/vehicle/index.ts
