+++ ppr-ui/src/components/tables/ppr/SearchedResultsPpr.vue
